+++ CHANGELOG.md
@@ -68,6 +68,8 @@ Changelog
  * removed `PackageState::NOT_LOADED`
  * removed `PackageState::detect()`
  * changed `DiscoveryManager::addBinding()` to accept a `BindingDescriptor` instance
+ * added `BindingDescriptor::listPathMappings()`
+ * added `BindingDescriptor::listRepositoryPaths()`
 
 * 1.0.0-beta (2015-01-12)